Changed Dni internal structure to improve performance
Before, the Dni digits were represented as a list of str. This made
updating a Dni missing digits an expensive operation.

The performance issues were not noticeable unless we were trying to
find_all_possible_dnis with more than 6 digits missing.

This change has reduce the time to find_all_possible_dnis with 8 digits
to half the time it took before